pyproject.toml
[tool.poetry]
name = "exercicioebac-djangopoetry"
version = "0.1.0"
description = "exercicio_EBAC_djangoPoetry"
authors = ["Dimas_Marcondes <dimasmarcondes88@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pytest = "^8.3.2"
factory-boy = "^3.3.0"
django = "^5.0.7"
django-extensions = "^3.2.3"
django-debug-toolbar = "^4.4.6"
faker = "^26.1.0"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
